feat: :sparkles: Changes to the codebase
Changes to the codebase:  - Updated the Docker image name from `tools_python:0.0.1` to `ander1221/tools_python:0.0.1`. - Changed the working directory in the Dockerfile from `/code` to `/app`. - Modified the `COPY` command in the Dockerfile to copy the entire project directory to `/app` instead of just the `requirements.txt` file. - Updated the `CMD` command in the Dockerfile to use `uvicorn` with the `--reload` flag. - Replaced `cv2` with `opencv-python` in the `requirements.txt` file.